Rebase gp_filedump to pg_filedump 12.0
This commit rebase gp_filedump [1] to pg_filedump [2], and make it work
for gpdb 4.3, 5, 6 and 7 (currently postgres 9.6 based).

The original gp_filedump master is at commit
29925bb81b9865a5ce140a6d3efbc127c163631f, it seems to be based on
pg_filedump-8.2 [3], but with many extra changes to comment styles and
indentations.  It is for gpdb 4.0 maybe, some hacks are needed to
compile on gpdb 4.3.

This patch is generated against pg_filedump master, commit
6a341df11148d5db7e5a04d5f0133635856579bd, which also has the tag
REL_12_0.

The patch is not a result of 'git rebase' or so, but actually some kind
of rewriting.  We first generated a diff from pg_filedump-8.2 to
gp_filedump, then we applied the diff manually.  We have to do so
because the code base has been changed a lot, both gpdb and pg_filedump.
During the process we refactored a lot.  We put some version checkings
to compile against all the current gpdb major versions.

Some options were retired, such are disabled with FIXMEs, here are a
list of notables:

- the program name is renamed from gp_filedump to pg_filedump;
- the -D option is retired: it specifies the datadir for the segment,
  however it has no effect even in gp_filedump, and a conflict -D exists
  in gp_filedump, which disables dead tuple info while deparsing;
- the -M option is retired: it is the checksum option for mirror files,
  it has no effect in gp_filedump either, there is a conflict -M,
  checksum for appendonly files, and in pg_filedump 12 there is already
  a -k option to verify checksums;
- the -p option is disabled with FIXME: it deparses heap data, I didn't
  find out the effect of it on gp_filedump, and in pg_filedump 12 there
  is a -D option to decode heap tuples, which works better, we could
  consider retire the -p option entirely;
- for heap tuple decoding, we need to add decoders for gpdb specific
  types;
- the -B and -D options are disabled with FIXME: they are "deparse the
  input block directory file" and "disable dead tuple info while
  deparsing", we could consider enable them later;

With current patch we could compile pg_filedump 12 against gpdb 4.3, 5,
6 and 7 (currently postgres 9.6 based).  Backward compatibilities are
provided so a build with gpdb 7 can also dump the heap/AO/CO files of
gpdb 4.3; however not the reverse, a build with gpdb 5 can not dump the
AO/CO files compressed in zstd, a build with gpdb 4.3 might show less
attributes and even with different attribute names. The pg_control file,
however, is version specific, so a build with gpdb 7 cannot dump the
pg_control file of gpdb 6.

Note For gpdb 4.3, under gpdb 4.3 source tree there is a workable
version under contrib/gp_filedump, personally I recommend it over the
current patch, the code base has changed too much, some gpdb specific
flags might not be handled in current patch.

[1]: https://github.com/pivotal/gp_filedump.git
[2]: https://git.postgresql.org/git/pg_filedump.git
[3]: http://mirrors.infowest.com/postgresql/projects/pgFoundry/pgfiledump/pg_filedump-8.2.tar.gz