Download gpdb source files instead of linking existing ones
To compile pg_filedump for gpdb we used to require a full local copy of
gpdb source tree, this is not always possible, so now we download the
needed source files from github directly.

The new build instructions are like below:

    # make sure we can find the gpdb package
    source /path/to/greenplum_path.sh

    # download needed gpdb source files according to the gpdb version
    make -f Makefile.gpdb download

    # build the pg_filedump
    make -f Makefile.gpdb

However we have to drop the gpdb 4.3 support because it is not open
sourced, an alternative way is to compile pg_filedump with gpdb 6 and
use it on gpdb 4.3 data files, or use the gp_filedump come with the gpdb
4.3 official package.

The quicklz support is also dropped.  We still keep the logic to support
quicklz as long as the quicklz library and headers can be found, but
this is usually not possible.