Split the tests for each GDPB major version
There is no point running all tests using a pg_filedump that's compiled against
a specific GPDB major version. Split them, so that only related ones are run
when we do installcheck.

Since there are just a handful of tests, there is no need to add schedules for
each one. Plus, we need to check if zlib/zstd/quicklz is enabled or not before
we could run the corresponding tests.