Add tests for new relation files instead of pre-generated ones
The new test cases are the same as existing ones, just that the relation files
are created using the running GPDB server, instead of pre-generated ones under
the test/ dirctory. The new cases are useful to catch errors caused by changes
in GPDB itself. This is part of the preparation effort to release pg_filedump
which includes running these tests in GPDB pipelines.

The "run_test.sql" script has been done in similar fashion as the upstream in
https://github.com/df7cb/pg_filedump/commit/b03319ddec99362c047729af3dc7e560712740e1, with GPDB-specific modifications.