Remove 4X/5X/6X tests from main
As we've made the main branch to only support the latest GPDB (current V7,
tracking Postgres version 12.12), now removing the tests that are unrelated.
Also, since GPDB7 doesn't support quicklz, remove that test as well.

The tests will be kept in the 'historical' branch for use for the older versions.